// packages/ui-inbox/src/settings/integrations/types.ts
export type IntegrationMutationVariables = {
  brandId: string;
  name: string;
  channelIds?: string[];
  visibility?: string;
  departmentIds?: string[];
  details?: any;
};